refactor, UI, build: move bootstrap css global and update imports
