+++ components/search/SearchResults.vue
@@ -61,7 +61,7 @@
 
 <script setup lang="ts">
 
-import "../assets/global_style1/bootstrap.min.css";
+import "../bootstrap_css/global_style1/bootstrap.min.css";
 import { ref, onMounted } from "vue";
 import type { PaperUI } from "~/types/research-paper-ui";
 const { getResearchPaper } = usePaper();
